// src/types/Country.ts
  export interface Province {
    code: string | number;
    name: string;
    [key: string]: any;
  }
  export interface District {
    code: string | number;
    name: string;
    [key: string]: any;
  }
  export interface Ward {
    code: string | number;
    name: string;
    [key: string]: any;
  }